fix bug with int, replace with uint64_t

+++ src/kmer-index/klcp.h
@@ -7,7 +7,7 @@
 #include "bitarray.h"
 
 typedef struct {
-  int seq_len;
+  uint64_t seq_len;
   bitarray_t* klcp;
   uint64_t* prev;
   uint64_t* next;
